LayoutTests/editing/selection/ios/selection-highlight-for-live-text-in-scrollable-area-expected.txt

Verifies that the highlight and handles show up in the correct place when selecting Live Text inside a subscrollable container

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Select text in the image above. This test requires WebKitTestRunner.

PASS selectionRects.length is 1
PASS UIHelper.rectContainsOtherRect(imageRect, selectionRects[0]) is true
PASS UIHelper.rectContainsOtherRect(imageRect, startHandleRect) is true
PASS UIHelper.rectContainsOtherRect(imageRect, endHandleRect) is true
PASS successfullyParsed is true

TEST COMPLETE
